refactor: route messages to relevant actors
