+++ src/server.rs
@@ -149,15 +149,15 @@ impl Server {
         naive_logging::log(&self.id, "running...");
 
         let cancellation_token = CancellationToken::new();
-        let follower_cancellation_token = cancellation_token.clone();
-        let candidate_cancellation_token = cancellation_token.clone();
+        let follower_ctk = cancellation_token.clone();
+        let candidate_ctk = cancellation_token.clone();
 
         let (follower_tx, follower_rx) = mpsc::channel(32);
         let (candidate_tx, candidate_rx) = mpsc::channel(32);
         tokio::join!(
             actors::run_root_actor(self, rx, follower_tx, candidate_tx),
-            actors::run_follower_actor(self, follower_rx, follower_cancellation_token),
-            actors::run_candidate_actor(self, candidate_rx, candidate_cancellation_token)
+            actors::run_follower_actor(self, follower_rx, follower_ctk),
+            actors::run_candidate_actor(self, candidate_rx, candidate_ctk)
         );
 
         Ok(self.id)
